requirements/development.txt
wheel==0.38.4

asgiref==3.5.2
boto3==1.26.12
bs4==0.0.1
celery==5.2.7
Collectfast==2.2.0
dj-database-url==1.0.0
Django==4.1.10
django-anymail==8.1
django-autoslug==1.9.8
django-axes==5.40.1
django-cors-headers==3.13.0
django_debug_toolbar==3.8.1
django-dirtyfields==1.9.0
django-dotenv==1.4.2
django-extensions==3.2.1
django-filter==22.1
django-health-check==3.17.0
django-ipware==4.0.2
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django-redis==5.2.0
django-redis-sessions==0.6.2
django-rest-passwordreset==1.3.0
django-storages==1.13.2
django-simple-history==3.2.0
django-versatileimagefield==2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
elasticsearch==8.5.0
firebase-admin==6.0.1
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-docx==0.8.11
pytz==2021.3
selenium==4.7.2
six==1.15.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
stripe==5.4.0
tqdm==4.65.0
twilio==7.16.4

